+++ crates/core/tests/java/case_statement_on_new_line/cases.out.java
@@ -1,0 +1,12 @@
+class A {
+    void m(int x) {
+        switch (x) {
+            case 1: foo();
+                bar();
+                break;
+            case 2: baz();
+                break;
+            default: break;
+        }
+    }
+}