+++ crates/core/tests/java/indent_break_from_case/cases.out.java
@@ -1,0 +1,16 @@
+class A {
+    void m(int x) {
+        while (true) {
+            switch (x) {
+                case 1:
+                    foo();
+                continue;
+                case 2:
+                    bar();
+                return;
+                default:
+                break;
+            }
+        }
+    }
+}